Create core class for ECS logic
